% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_leagues.R, R/available_seasons_.R
\name{available_seasons}
\alias{available_seasons}
\title{Print available football seasons}
\usage{
available_seasons()

available_seasons()
}
\value{
Prints the available football seasons to the console.

Prints the available football seasons to the console.
}
\description{
This function prints the football seasons for which statistics are available.

This function prints the available football seasons for which statistics
are available.
}
\examples{

# Print available football seasons
available_seasons()



# Print available football seasons
available_seasons()

}
