% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_big_5_player_stats.R
\name{fbref_big5_player_stats}
\alias{fbref_big5_player_stats}
\title{Get football statistics for players from the top 5 European leagues (Big 5).}
\arguments{
\item{season}{The season for which statistics are requested
(e.g., "2022/2023"). If not provided, the most recent season will be used.}

\item{type}{The type of statistics to retrieve (e.g., "standard",
"goalkeeping", "shooting"). If not provided, "standard" statistics
will be used.}
}
\value{
A data frame containing football statistics for players in the
specified season and of the specified type.
}
\description{
This function retrieves football statistics for players from the top 5
European leagues, commonly referred to as the "Big 5" leagues.
}
\details{
These leagues include:
- English Premier League (EPL)
- La Liga (Spain)
- Serie A (Italy)
- Bundesliga (Germany)
- Ligue 1 (France)
}
\examples{
\dontrun{
# Get passing player statistics for the 2022/2023 season
fbref_big5_player_stats(season = "2022/2023", type = "passing")
}

}
\seealso{
\url{https://fbref.com/en/comps/Big5/Big-5-European-Leagues-Stats}
}
