% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_br_serie_a_team_stats.R
\name{fbref_br_serie_a_team_stats}
\alias{fbref_br_serie_a_team_stats}
\title{Scrape Campeonato Brasileiro Série A Team Statistics from FBref}
\arguments{
\item{season}{A string specifying the season in the format "YYYY/YYYY" or
"YY/YY". If not provided, it defaults to the current season.}

\item{type}{A string specifying the type of statistics to retrieve (e.g.,
"standard", "goalkeeping", "shooting"). If not provided, it defaults
to "standard".}
}
\value{
A data frame containing the scraped team statistics.
}
\description{
This function scrapes team statistics for Serie A from FBref based on the
season and type provided.
}
\examples{
\dontrun{
# Série A season
fbref_br_serie_a_team_stats()
}

}
\seealso{
\url{https://fbref.com/en/comps/24/Serie-A-Stats}
}
