% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_bundesliga_team_stats.R
\name{fbref_bundesliga_team_stats}
\alias{fbref_bundesliga_team_stats}
\title{Scrape German Bundesliga Team Statistics from FBref}
\arguments{
\item{season}{A string specifying the season in the format "YYYY/YYYY" or
"YY/YY".}

\item{type}{A string specifying the type of statistics to retrieve
(e.g., "standard", "goalkeeping", "shooting").}
}
\value{
A data frame containing the scraped team statistics.
}
\description{
This function scrapes team statistics for the Bundesliga from FBref based
on the season and type provided.
}
\examples{
\dontrun{
fbref_bundesliga_team_stats(season = "2019/2020", type = "shooting")
}

}
\seealso{
\url{https://fbref.com/en/comps/20/Bundesliga-Stats}
}
