% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_efl_team_stats.R
\name{fbref_efl_team_stats}
\alias{fbref_efl_team_stats}
\title{Get EFL Championship team statistics from FBref}
\arguments{
\item{season}{A character string representing the season in the
format 'YYYY/YYYY'.}

\item{type}{A character string specifying the type of statistics to retrieve.
Valid types are: 'standard', 'goalkeeping', 'advanced_goalkeeping',
'shooting', 'passing',
'pass_types', 'goal_creation', 'defensive_actions', 'possession',
'playing_time', or 'miscellaneous'.}
}
\value{
A data frame containing the requested team statistics.
}
\description{
This function retrieves EFL Championship team statistics from FBref for a
specific season and type.
}
\details{
The function uses the 'rvest' package to scrape data from FBref.
The 'season' parameter should be in a valid 'YYYY/YYYY' format.
}
\examples{
fbref_efl_team_stats("2021/2022", "standard")

}
\seealso{
\url{https://fbref.com/en/comps/10/Championship-Stats}
}
