% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_epl_team_stats.R
\name{fbref_epl_team_stats}
\alias{fbref_epl_team_stats}
\title{Retrieve English Premier League team statistics from FBref.}
\arguments{
\item{season}{A character string representing the season (e.g., "2022/2023").
If NULL, the current season is used.}

\item{type}{A character string representing the type of statistics to
retrieve (e.g., "standard"). Default is "standard."}
}
\value{
A data frame containing the requested team statistics.
}
\description{
This function scrapes team statistics for the English Premier League from FBref.com.
}
\details{
This function fetches English Premier League team statistics from
FBref based on the provided season and type.
}
\examples{
\dontrun{
fbref_epl_team_stats(season = "2022/2023", type = "standard")
}

}
\seealso{
\url{https://fbref.com/en/comps/9/Premier-League-Stats}
}
