% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_eredivisie_team_stats.R
\name{fbref_eredivisie_team_stats}
\alias{fbref_eredivisie_team_stats}
\title{Scrape Eredivisie team statistics from FBref}
\arguments{
\item{season}{A character vector specifying the season for which you want to
retrieve statistics. The season should be in the format "YYYY/YYYY" or
"YY/YY" (e.g., "2022/2023" or "22/23"). If not provided, the function will
default to the most recent season available.}

\item{type}{A character vector specifying the type of statistics to retrieve.
Available types are "standard," "goalkeeping," "advanced_goalkeeping,"
"shooting," "passing," "pass_types," "goal_creation," "defensive_actions,"
"possession," "playing_time," or "miscellaneous." If not provided, the
function will default to "standard."}
}
\value{
A data frame containing the requested team statistics for the
specified season and type.
}
\description{
This function scrapes team statistics for the Eredivisie from FBref.com.
}
\details{
This function scrapes team statistics for the Eredivisie football
league from FBref.com, a reputable source for football statistics.
The function allows you to specify the season and the type of statistics you
want to retrieve.
}
\examples{
# Retrieve standard team statistics for the 2022/2023 season
fbref_eredivisie_team_stats(season = "2022/2023", type = "standard")

# Retrieve goalkeeping team statistics for the current season
fbref_eredivisie_team_stats(type = "goalkeeping")

# Retrieve advanced goalkeeping team statistics for the 2021/2022 season
fbref_eredivisie_team_stats(season = "2021/2022", type = "advanced_goalkeeping")

}
\seealso{
\url{https://fbref.com/en/comps/23/Eredivisie-Stats}

\url{https://fbref.com/en/comps/23/Eredivisie-Stats}
}
