% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_liga_mx_team_stats.R
\name{fbref_liga_mx_team_stats}
\alias{fbref_liga_mx_team_stats}
\title{Scrape Liga MX team statistics from FBRef.com}
\arguments{
\item{season}{The season to retrieve statistics for (e.g., "2022/2023").
If not provided, the current season will be used.}

\item{type}{The type of statistics to retrieve (e.g., "standard"). If not
provided, "standard" will be used.}
}
\value{
A data frame containing the scraped team statistics.
}
\description{
This function scrapes team statistics for Liga MX from FBRef.com for a
specific season and data type.
}
\examples{
fbref_liga_mx_team_stats(season = "2022/2023", type = "standard")

}
\seealso{
\url{https://fbref.com/en/comps/31/Liga-MX-Stats}
}
