% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_ligue1_team_stats.R
\name{fbref_ligue1_team_stats}
\alias{fbref_ligue1_team_stats}
\title{Scrape French Ligue 1 Team Statistics from FBref}
\arguments{
\item{season}{The season you want to retrieve statistics for. If not provided, the function
defaults to the current season.}

\item{type}{The type of statistics to retrieve. Options include "standard," "goalkeeping,"
"advanced_goalkeeping," "shooting," "passing," "pass_types," "goal_creation,"
"defensive_actions," "possession," "playing_time," and "miscellaneous."}
}
\value{
A data frame containing the requested team statistics for the specified season and type.
}
\description{
This function scrapes team statistics for Ligue 1 from FBref.com.
}
\examples{
# Scrape Standard Team Statistics for the 2021/2022 season
fbref_ligue1_team_stats(season = "2021/2022", type = "standard")

# Scrape possession Team Statistics for the 2022/2023 season
fbref_ligue1_team_stats(season = "2022/2023", type = "possession")

# Scrape Pass types Team Statistics for the current season (default)
fbref_ligue1_team_stats(type = "pass_types")


}
\seealso{
\url{https://fbref.com/en/comps/13/Ligue-2-Stats}
}
