% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_mls_team_stats.R
\name{fbref_mls_team_stats}
\alias{fbref_mls_team_stats}
\title{Retrieve Major League Soccer (MLS) team statistics from FBref.}
\arguments{
\item{season}{The season for which you want to retrieve statistics.
Should be in the format 'YYYY' (e.g., '2023').
              If not provided, the current season will be used.}

\item{type}{The type of statistics to retrieve. Available types include
 'standard', 'goalkeeping', 'advanced_goalkeeping',
'shooting', 'passing', 'pass_types', 'goal_creation',
'defensive_actions', 'possession', 'playing_time',
and 'miscellaneous'. If not provided, 'standard' statistics will
be retrieved.}
}
\value{
A data frame containing the requested MLS team statistics.
}
\description{
This function scrapes team statistics for Major League Soccer (MLS) from
FBref for a specified season and type.
}
\examples{
# Retrieve standard MLS team statistics for the 2022 season
fbref_mls_team_stats(season = '2022', type = 'standard')

# Retrieve goalkeeping statistics for the current MLS season
fbref_mls_team_stats(type = 'goalkeeping')

}
\seealso{
\url{https://fbref.com/en/comps/22/Major-League-Soccer-Stats}
}
