% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_similar_players.R
\name{find_similar_players}
\alias{find_similar_players}
\title{Retrieve football team statistics from FBREF}
\usage{
find_similar_players(
  df,
  player,
  metrics = NULL,
  formula = "cosine",
  top_n = 15
)
}
\arguments{
\item{df}{The data scraped using kickR's fbref_player_stats() or
fbref_big_5_player_stats()}

\item{player}{The player(target) for whom you would like to find other
similar players}

\item{metrics}{The metrics you would like to find similar players for. For
example if you want to find similar players in possession, a metric that
would interest you might be inplay touches/passes etc. The function would
return players, whose output in said metric(s) are similar to the target
player supplied.}

\item{formula}{The type of formula to calculate similarity. This function
uses cosine or euclidean but defaults to cosine similarity. For a metric
or smaller group of similar metrics, cosine performs better.}

\item{top_n}{The number of similar players you want to return.
function defaults to top 15 most similar players.}
}
\value{
A data frame containing similar players based on cosine similarity
  or euclidean distance.
}
\description{
This function retrieves football team statistics from FBREF for a specified
league, season, and type of statistics.
}
\examples{
\dontrun{
# Find similar players to Martin Ødegaard in terms of specified passing metrics
find_similar_players(df = df,
                     player ="Martin Ødegaard",
                     metrics = c("xAG","xA","key_passes",
                                 "passes_into_final_third",
                                 "passes_into_penalty_box",
                                 "progressive_passes"),
                     formula = "euclidean",
                     top_n = 15)
}
}
