% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_table.R
\name{save_table}
\alias{save_table}
\title{Save a DataFrame in Multiple Formats}
\arguments{
\item{df}{The DataFrame to be saved.}

\item{filename}{The name (excluding file extension) for the saved file.}

\item{format}{A character vector specifying the formats to save the DataFrame in. The available options are "json", "csv", "xlsx", and "rds".}
}
\value{
This function saves the DataFrame in the specified formats.
}
\description{
This function allows you to save a DataFrame in various formats such as
JSON, CSV, XLSX, and RDS.
}
