% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_types.R
\name{stat_types}
\alias{stat_types}
\title{Print available football team statistics types}
\usage{
stat_types()
}
\value{
Prints the available statistics types to the console.
}
\description{
This function prints the available statistics types that are available for leagues.
}
\examples{

# Print available statistics types
stat_types()


}
