% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_bundesliga_team_stats.R
\name{fbref_bundesliga_team_stats}
\alias{fbref_bundesliga_team_stats}
\title{Scrape German Bundesliga Team Statistics from FBref}
\arguments{
\item{season}{A string specifying the season in the format "YYYY/YYYY" or
"YY/YY".}

\item{type}{A string specifying the type of statistics to retrieve
(e.g., "standard", "goalkeeping", "shooting").}
}
\value{
A data frame containing the scraped team statistics.
}
\description{
This function scrapes team statistics for the Bundesliga from FBref based
on the season and type provided.
}
\examples{
# Scrape standard team statistics for the 2022/2023 Bundesliga season
bundesliga_stats_standard_2022 <- fbref_bundesliga_team_stats(
          season = "2022/2023", type = "standard")

# Scrape goalkeeping team statistics for the 2021/2022 Bundesliga season
bundesliga_stats_goalkeeping_2021 <- fbref_bundesliga_team_stats(
          season = "2021/2022", type = "goalkeeping")

# Scrape shooting team statistics for the 2019/2020 Bundesliga season
bundesliga_stats_shooting_2019 <- fbref_bundesliga_team_stats(
          season = "2019/2020", type = "shooting")

# Scrape possession team statistics for the current Bundesliga season
bundesliga_stats_possession_current <- fbref_bundesliga_team_stats(
          type = "possession")

}
\seealso{
\url{https://fbref.com/en/comps/20/Bundesliga-Stats}
}
